chore: refactor scheduled jobs to use redisTemplate based locks
